requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
mlxtend>=0.18.0
mysql-connector-python>=8.0.0
openpyxl>=3.6.0
